+++ NEWS.md
@@ -1,7 +1,10 @@
 # osmapiR (development version)
 
 * Use the new function `httr2::oauth_cache_clear()` from httr2 1.0.6 (#58 by @hadley).
-* Update documentation and code for server-side changes documented in OSMWikiVersion 2775892 (#60).
+* Update documentation and code for server-side changes documented in OSMWikiVersion
+  [2711808 -> 2775892](https://wiki.openstreetmap.org/w/index.php?title=API_v0.6&type=revision&diff=2775892&oldid=2711808)
+  (#60).
+  * Add new parameters to `osm_query_changesets(..., from, to)`.
 
 # osmapiR 0.2.1